graphiti-main/claude_docker/requirements.txt
# Claude Docker API Requirements
# These are the minimal requirements for running the Claude Docker API server

# API Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
httpx==0.27.2

# Async support
aiofiles==24.1.0
asyncio==3.4.3

# For running tests
pytest==8.3.3
pytest-asyncio==0.24.0

# Optional: For development
python-dotenv==1.0.1